pyproject.toml
[build-system]
requires = ["setuptools>=79,<81", "wheel~=0.45", "setuptools_scm[toml]>=8.3,<10.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ixbrl-viewer"
dynamic = ['version']
readme = "README.md"
license = "Apache-2.0"
description = "The Arelle iXBRL Viewer allows iXBRL reports to be viewed interactively in a web browser."
authors = [
    {name = "arelle.org", email = "support@arelle.org"}
]
keywords = ["arelle", "plugin", "xbrl"]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: End Users/Desktop',
    'Intended Audience :: Developers',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Programming Language :: Python :: 3.14',
    'Operating System :: OS Independent',
    'Topic :: Text Processing :: Markup :: XML'
]
requires-python = ">=3.10"
dependencies = [
    'bottle>=0.13,<0.14',
    'lxml>=4,<7'
]
[project.optional-dependencies]
arelle = [
    'arelle_release==2.*',
]
dev = [
    'flake8==7.3.0',
    'mypy==1.19.0',
    'pytest==9.0.2',
    'types-lxml==2025.11.25',
    'typing-extensions==4.15.0',
]

# https://setuptools.pypa.io/en/latest/userguide/entry_point.html#entry-points-for-plugins
[project.entry-points."arelle.plugin"]
ixbrl-viewer = "iXBRLViewerPlugin:load_plugin_url"

[project.urls]
"Homepage" = "https://arelle.org/"
"Downloads" = "https://arelle.org/arelle/pub/"
"Documentation" = "https://arelle.org/arelle/documentation/"
"Blog" = "https://arelle.org/arelle/blog/"
"Source Code" = "https://github.com/Arelle/ixbrl-viewer"
"Bug Reports" = "https://github.com/Arelle/ixbrl-viewer/issues/"
"Support" = "https://groups.google.com/g/arelle-users/"

[tool.setuptools]
platforms = ["any"]

[tool.setuptools.packages.find]
include = ["iXBRLViewerPlugin*"]

[tool.setuptools_scm]
tag_regex = "^(?:[\\w-]+-?)?(?P<version>[vV]?\\d+(?:\\.\\d+){0,2}[^\\+]*)(?:\\+.*)?$"
write_to = "iXBRLViewerPlugin/_version.py"

[tool.mypy]
# Warn when a # type: ignore comment does not specify any error codes
enable_error_code = "ignore-without-code"
python_version = "3.14"
show_error_codes = true
strict = true

# --- arelle types ---
# We don't publish arelle types (yet), as they're incomplete and we don't want other devs to rely on them.
# But we should be using them ourselves to gradually improve typing in Arelle and its plugins.
[[tool.mypy.overrides]]
module = 'arelle.*'
follow_untyped_imports = true
